+++ WebApiTest/Services/HusbandService/IHusbandService.cs
@@ -10,7 +10,7 @@ namespace WebApiTest.Services.HusbandService
         public Task<List<WantedList>> GetWantedListAsync();
         public Task<List<Shop>> GetShopsForVisitAsync();
         public Task<List<Product>> GetProductsInShopAsync(int ShopId);
-        public Task<Shop> FindShopAsync(long id);
+        public Task<Shop> FindShopAsync(int id);
         public Task<Product> FindProductAsync(string name);
     }
 }
